Removed default component files, added a whole bunch of classes.
When you generate a plugin with atlassian cli it creates some files for you, I Just haven't removed them until now.
SEPCommits, SEPCommitsImpl are the interface/class for munging all the commit data out of refChange event.
SEPRefChangeEvent and SEPRefChangeEventImpl are the interface/class for handing the refChange event and kicking off all
subevents that occur because of a refchange.