Fixed an issue where I was opening too many relay connections.
My plugin is set up to handle a variety of types of ref change events; stuff like branch/commit/tag creation, branch/commit/tag deletion, etc. Currently, all I really support is commit creation and deletion but it opens a connection to the relay for all of these types of events regardless. However, the only time it closes the connection is if it were a commit creation or deletion. This means that whenever someone  created a tag or branch or deleted them, a connection to the relay would open and never be closed. I fixed it by doing something very simple: I open connection to the relay, process the event, and then close the connection all in the same method. This way, it’s agnostic about what kind of event it is.