Logger fixes
Setting additivity to false means that the appender I specify in this config will only be attached to the logger that I specify.
This means that we wont be getting duplicate messages in both Stash's main log and this one. I also needed to change the location
of the file because that variable (log.dir) isn't available to me.